# docs/openapi.yaml
openapi: "3.0.2"
info:
  title: "Vibe - Microblog"
  version: "0.1.0"
paths:
  /register:
    post:
      summary: "Register a new user"
      description: "Create a new user account and return an access token."
      operationId: "register_register_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /token:
    post:
      summary: "User login (OAuth2)"
      description: "Obtain an access token using username and password (form-data). Use returned bearer token for authenticated endpoints."
      operationId: "login_token_post"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Body_login_token_post"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /posts:
    post:
      summary: "Create a post"
      description: "Create a short text post (reply via `parent_id` for one-level replies). Requires Bearer token."
      operationId: "create_post_posts_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCreate"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostOut"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          OAuth2PasswordBearer:
  /feed:
    get:
      summary: "Global feed"
      description: "Return a chronological global feed. Supports pagination via `page` (0-indexed) and `page_size`."
      operationId: "feed_feed_get"
      parameters:
        -
          required: false
          schema:
            title: "Page"
            type: "integer"
            default: 0
          name: "page"
          in: "query"
        -
          required: false
          schema:
            title: "Page Size"
            type: "integer"
            default: 50
          name: "page_size"
          in: "query"
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                title: "Response Feed Feed Get"
                type: "array"
                items:
                  $ref: "#/components/schemas/PostOut"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /posts/{post_id}/like:
    post:
      summary: "Like a post"
      description: "Like a post by id. Requires Bearer token. Duplicate likes are rejected."
      operationId: "like_post_posts__post_id__like_post"
      parameters:
        -
          required: true
          schema:
            title: "Post Id"
            type: "integer"
          name: "post_id"
          in: "path"
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          OAuth2PasswordBearer:
  /users/{username}:
    get:
      summary: "User profile"
      description: "View a user's profile and their posts in chronological order."
      operationId: "profile_users__username__get"
      parameters:
        -
          required: true
          schema:
            title: "Username"
            type: "string"
          name: "username"
          in: "path"
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileOut"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    Body_login_token_post:
      title: "Body_login_token_post"
      required:
        - "username"
        - "password"
      type: "object"
      properties:
        grant_type:
          title: "Grant Type"
          pattern: "password"
          type: "string"
        username:
          title: "Username"
          type: "string"
        password:
          title: "Password"
          type: "string"
        scope:
          title: "Scope"
          type: "string"
          default: ""
        client_id:
          title: "Client Id"
          type: "string"
        client_secret:
          title: "Client Secret"
          type: "string"
    HTTPValidationError:
      title: "HTTPValidationError"
      type: "object"
      properties:
        detail:
          title: "Detail"
          type: "array"
          items:
            $ref: "#/components/schemas/ValidationError"
    PostCreate:
      title: "PostCreate"
      required:
        - "content"
      type: "object"
      properties:
        content:
          title: "Content"
          maxLength: 280
          minLength: 1
          type: "string"
        parent_id:
          title: "Parent Id"
          type: "integer"
    PostOut:
      title: "PostOut"
      required:
        - "id"
        - "author_id"
        - "content"
        - "created_at"
        - "likes"
      type: "object"
      properties:
        id:
          title: "Id"
          type: "integer"
        author_id:
          title: "Author Id"
          type: "integer"
        content:
          title: "Content"
          type: "string"
        created_at:
          title: "Created At"
          type: "string"
          format: "date-time"
        parent_id:
          title: "Parent Id"
          type: "integer"
        likes:
          title: "Likes"
          type: "integer"
        replies:
          title: "Replies"
          type: "array"
          items:
            $ref: "#/components/schemas/PostOut"
          default:
    ProfileOut:
      title: "ProfileOut"
      required:
        - "id"
        - "username"
        - "created_at"
      type: "object"
      properties:
        id:
          title: "Id"
          type: "integer"
        username:
          title: "Username"
          type: "string"
        display_name:
          title: "Display Name"
          type: "string"
        created_at:
          title: "Created At"
          type: "string"
          format: "date-time"
        posts:
          title: "Posts"
          type: "array"
          items:
            $ref: "#/components/schemas/PostOut"
          default:
    Token:
      title: "Token"
      required:
        - "access_token"
      type: "object"
      properties:
        access_token:
          title: "Access Token"
          type: "string"
        token_type:
          title: "Token Type"
          type: "string"
          default: "bearer"
    UserCreate:
      title: "UserCreate"
      required:
        - "username"
        - "password"
      type: "object"
      properties:
        username:
          title: "Username"
          maxLength: 32
          minLength: 3
          type: "string"
        password:
          title: "Password"
          minLength: 6
          type: "string"
        display_name:
          title: "Display Name"
          type: "string"
    ValidationError:
      title: "ValidationError"
      required:
        - "loc"
        - "msg"
        - "type"
      type: "object"
      properties:
        loc:
          title: "Location"
          type: "array"
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
        msg:
          title: "Message"
          type: "string"
        type:
          title: "Error Type"
          type: "string"
  securitySchemes:
    OAuth2PasswordBearer:
      type: "oauth2"
      flows:
        password:
          scopes:
          tokenUrl: "/token"